Move ViewRouter to /src/layouts folder
